+++ src/Phan/Config.php
@@ -61,6 +61,9 @@ class Config
     private static $array_casts_as_null = false;
 
     /** @var bool replicates configuration with the same name */
+    private static $strict_method_checking = false;
+
+    /** @var bool replicates configuration with the same name */
     private static $strict_param_checking = false;
 
     /** @var bool replicates configuration with the same name */
@@ -291,6 +294,12 @@ class Config
         // type can be cast to null. Setting this to true
         // will cut down on false positives.
         'null_casts_as_any_type' => false,
+
+        // If enabled, Phan will warn if **any** type in the method's object
+        // is definitely not an object.
+        // Setting this to true will introduce numerous false positives
+        // (and reveal some bugs).
+        'strict_method_checking' => false,
 
         // If enabled, Phan will warn if **any** type in the argument's type
         // cannot be cast to a type in the parameter's expected type.
@@ -953,6 +962,11 @@ class Config
         return self::$null_casts_as_any_type;
     }
 
+    public static function get_strict_method_checking() : bool
+    {
+        return self::$strict_method_checking;
+    }
+
     public static function get_strict_param_checking() : bool
     {
         return self::$strict_param_checking;
@@ -1036,6 +1050,9 @@ class Config
                 break;
             case 'array_casts_as_null':
                 self::$array_casts_as_null = $value;
+                break;
+            case 'strict_method_checking':
+                self::$strict_method_checking = $value;
                 break;
             case 'strict_param_checking':
                 self::$strict_param_checking = $value;