+++ src/Phan/Language/EmptyUnionType.php
@@ -1124,4 +1124,14 @@ final class EmptyUnionType extends UnionType
     {
         return null;
     }
+
+    /**
+     * Returns true if this contains a type that is definitely nullable or a non-object.
+     * e.g. returns true for ?T, T|false, T|array
+     *      returns false for T|callable, object, T|iterable, etc.
+     */
+    public function containsDefiniteNonObjectType() : bool
+    {
+        return false;
+    }
 }