+++ src/Phan/Language/Type.php
@@ -2629,4 +2629,14 @@ class Type
     {
         return false;
     }
+
+    /**
+     * Returns true if this contains a type that is definitely nullable or a non-object.
+     * e.g. returns true false, array, int
+     *      returns false for callable, object, iterable, T, etc.
+     */
+    public function isDefiniteNonObjectType() : bool
+    {
+        return false;
+    }
 }